requirements.txt
allure-pytest==2.9.45
pytest==7.1.2
selenium==4.1.3
pytest-html==3.1.1
pytest-xdist==2.5.0
pylint==2.13.9
isort==5.10.1
